many fixes, most op are OK

+++ VM/src/fork_op.c
@@ -6,7 +6,7 @@
 /*   By: akopera <akopera@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/01/29 18:26:17 by akopera           #+#    #+#             */
-/*   Updated: 2017/02/17 20:21:25 by akopera          ###   ########.fr       */
+/*   Updated: 2017/02/20 19:22:12 by tdefresn         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,5 +14,5 @@
 
 void	fork_op(t_proc *proc, t_op_arg args[3])
 {
-	process_fork(proc, args[0].value % IDX_MOD);
+	process_fork(proc, (short)args[0].value % IDX_MOD);
 }
