+++ VM/src/lfork_op.c
@@ -6,7 +6,7 @@
 /*   By: akopera <akopera@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/01/29 18:27:40 by akopera           #+#    #+#             */
-/*   Updated: 2017/02/17 20:18:55 by akopera          ###   ########.fr       */
+/*   Updated: 2017/02/20 19:21:48 by tdefresn         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,5 +14,5 @@
 
 void	lfork_op(t_proc *proc, t_op_arg args[3])
 {
-	process_fork(proc, args[0].value);
+	process_fork(proc, (short)args[0].value);
 }
